refactor: restructure layout and provider logic
